(Update) UI Logs
Mejoras en Ui si no hay logs, evento de borrar log creado
